refactor: 폴더 구조 변경 및 절대 경로 적용
