+++ src/stories/Header/Header.stories.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Header from "../../components/common/Header";
+import Header from "components/layout/Header";
 
 export default {
   title: "Component/Header/Header",
