+++ src/stories/Header/Title.stories.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Title from "../../components/common/Header/Title";
+import Title from "components/layout/Header/Title";
 
 export default {
   title: "Component/Header/Title",
