+++ src/stories/Spinner.stories.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Spinner from "../components/common/Spinner";
+import Spinner from "components/common/Spinner";
 
 export default {
   title: "Component/Common/Spinner",
